Lint

+++ src/main/java/lang/sep3asm/TokenAssoc.java
@@ -7,7 +7,7 @@ public class TokenAssoc {
 	private Object info;
 	public int getType()						{ return type; }
 	public Sep3Instruction getInstruction()		{ return (Sep3Instruction) info; }
-	public int getRegisterNumber()				{ return ((Integer) info).intValue(); }
+	public int getRegisterNumber()				{ return (Integer) info; }
 	public TokenAssoc(int type, Object info) {
 		this.type = type;
 		this.info = info;
